% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extantify.R
\name{Extantify}
\alias{Extantify}
\title{extantify}
\usage{
Extantify(subsetcolumns = NA, factors = NA, dir = "a")
}
\arguments{
\item{subsetcolumns}{(a vector of columns to subset based off of)}

\item{factors}{(Indicate what factors to select for the corresponding subset columns. Note this should be a vector or list of same length as subsetcolumns)}
}
\value{
extant (Extant fish species for a given subset)
}
\description{
subsetcolumns and extantifies the IUCN long data table based on desired output
}
\author{
Niko Hartline, Fish-ecol.github.io
}

