% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_index.R
\name{iucn_index}
\alias{iucn_index}
\title{iucn_index}
\usage{
iucn_index(data)
}
\arguments{
\item{data}{(vector of categories by country)}
}
\value{
iucn_index (The IUCN index)
}
\description{
Calculates a weighted average for IUCN's categories
}
\author{
Juan, Patricia, JC
}

